% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_conditional_ra.R
\name{conduct_conditional_ra}
\alias{conduct_conditional_ra}
\title{Conduct Conditional Random Assignment}
\usage{
conduct_conditional_ra(declaration, assignment_vec, conditions)
}
\arguments{
\item{declaration}{A random assignment declaration, created by \code{\link{declare_ra}}.}

\item{assignment_vec}{A vector of length N indicating the treatment conditions to which units were assigned.}

\item{conditions}{A character vector of length 2 indicating which treatment conditions are to be permuted.}
}
\description{
Conditional random assignment is useful when conducting randomization inference in multi-arm trials. When conducting a hypothesis test under a sharp null hypothesis for a *pairwise* comparison, we want to consider only random assignments in which units assigned to *other* conditions remain in their original condidtions.
}
\details{
For example, in a three-arm trial in which we want to conduct a hypothesis test under the sharp null hypothesis of exactly no difference in potential outcomes between conditions T1 and T2, we want to permute the random assignment for units that were assigned to T1 and T2, but we want to leave units assigned to T3 alone. We do this because our sharp null hypothesis does not say anything about potential outcomes under T3, so we do not want to have to imagine what the potential outcomes for units assigned to T3 *would have been* under alternative assignments.
}
