% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri.R
\name{conduct_ri_f}
\alias{conduct_ri_f}
\title{F stat}
\usage{
conduct_ri_f(model_1, model_2, assignment = "Z", declaration, data,
  IPW = TRUE, sharp_hypothesis = 0, sims = 1000)
}
\arguments{
\item{sims}{}
}
\description{
F stat
}
