% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri2.R
\docType{package}
\name{ri2-package}
\alias{ri2}
\alias{ri2-package}
\title{ri2: Randomization Inference}
\description{
Tools for conducting randomization inference for experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alexandercoppock.com/ri2/}
  \item \url{https://github.com/acoppock/ri2}
  \item Report bugs at \url{https://github.com/acoppock/ri2/issues}
}

}
\author{
\strong{Maintainer}: Alexander Coppock \email{acoppock@gmail.com}

Authors:
\itemize{
  \item Alexander Coppock \email{acoppock@gmail.com}
}

}
\keyword{internal}
